change buffer_size calculation, move pool.py to pgl library
